Save hive coordinates on drag end

+++ server/apiculture/models/core.py
@@ -6,7 +6,7 @@ from typing import List, Optional
 from sqlalchemy import ForeignKey, String
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column, relationship
 
-from apiculture.models.enum import HiveModels, HiveTypes, ApiaryTypes
+from apiculture.models.enum import ApiaryTypes, HiveModels, HiveTypes
 
 
 class Base(DeclarativeBase):
@@ -31,8 +31,7 @@ class Apiary(Base):
     owner_id: Mapped[int] = mapped_column(ForeignKey("user.id"))
     number: Mapped[str]
     name: Mapped[str]
-    type: Mapped[Optional[int]] = mapped_column(
-        default=ApiaryTypes.IMMOBILE.value)
+    type: Mapped[Optional[int]] = mapped_column(default=ApiaryTypes.IMMOBILE.value)
     hives: Mapped[Optional[List["Hive"]]] = relationship()
 
 
@@ -42,13 +41,11 @@ class Hive(Base):
     id: Mapped[int] = mapped_column(primary_key=True, autoincrement=True)
     number: Mapped[str]
     apiary_id: Mapped[int] = mapped_column(ForeignKey("apiary.id"))
-    model: Mapped[Optional[int]] = mapped_column(
-        default=HiveModels.DADAN_BLAT.value)
-    type: Mapped[Optional[int]] = mapped_column(
-        default=HiveTypes.BEE_FAMILY.value)
+    model: Mapped[Optional[int]] = mapped_column(default=HiveModels.DADAN_BLAT.value)
+    type: Mapped[Optional[int]] = mapped_column(default=HiveTypes.BEE_FAMILY.value)
     status: Mapped[Optional[str]]
-    x: Mapped[Optional[int]] = mapped_column(default=0)
-    y: Mapped[Optional[int]] = mapped_column(default=0)
+    x: Mapped[Optional[float]] = mapped_column(default=0.0)
+    y: Mapped[Optional[float]] = mapped_column(default=0.0)
 
     class Meta:
         unique_together = ("apiary_id", "number")
